backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.22
httpx==0.27.0
pydantic==2.8.0
python-dotenv==1.0.1
